chore(appium): refactor wdio config files
