+++ tests/specs/reusable-accounts/01-create-accounts-and-friends.spec.ts
@@ -170,7 +170,7 @@ export default async function createChatAccountsTests() {
     await expect(
       chatsLayoutFirstUser.encryptedMessagesText
     ).toHaveTextContaining(
-      "Messages are secured by end-to-end encryption and sent over a peer-to-peer network."
+      "MessageZZZ are secured by end-to-end encryption and sent over a peer-to-peer network."
     );
   });
